Final code update
